requirements.txt
flask==2.3.3
flask-cors==4.0.0
praw==7.7.1
nltk==3.8.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2